Autonomous-Navigation-Robot/requirements.txt
flask>=3.0.0
numpy>=1.24.0
opencv-python>=4.8.0
RPi.GPIO>=0.7.1; platform_machine=="armv7l"
